Change unoptimized-images feature policy to unoptimized-lossy-images.
This makes it more clear that this policy applies only to lossy image
types (for now, just JPEG). This patch also makes the implementation
more generic in order to make the policy compatible with parsed
policy values and to pave the way for implementing a
lossless-enforcing version of the unoptimized-images policy as well.

These changes have been pulled from a larger partially reviewed CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1524869

Bug: 942659
Change-Id: I0242ac900b1653b89d6d8dbd63616003dcaaba62